refactor(Promised): clean api
