--- plugins/PromisedScene/types/sites.ts
@@ -1,12 +1,0 @@
-export interface Data {
-  id: number;
-  name: string;
-  short_name: string;
-  url: string;
-  logo: string;
-  favicon: string;
-}
-
-export interface SiteListResult {
-  data: Data[];
-}
